pyproject.toml
[project]
name = "ac-server-manager"
version = "0.1.0"
description = "Automated AWS deployment for Assetto Corsa dedicated servers"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "boto3>=1.28.0",
    "click>=8.1.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "moto>=4.2.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "boto3-stubs[ec2,s3]>=1.28.0",
]

[project.scripts]
ac-deploy = "ac_server_manager.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=ac_server_manager --cov-report=term-missing"

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
strict = true
warn_return_any = true
warn_unused_configs = true
